Implement STATUS_RESTRICTED for resources
STATUS_RESTRICTED is needed in applications that make resources accessible to external users that don't have full acess to the application but limited access to an area outside of the application. For example a customer or partner web site. The application is responsible for protecting unauthorized access to RESTRICTED resources.